Move common classes for jira links
