--- src/coco/zeplinComponent/model/BarrelDetailsResponse.ts
@@ -1,9 +1,0 @@
-import ZeplinComponentSection from "../../../common/domain/zeplinComponent/model/ZeplinComponentSection";
-import ResponseBarrel from "../../barrel/model/ResponseBarrel";
-import ResponseScreenSection from "../../../sidebar/screen/model/ResponseScreenSection";
-
-export default interface BarrelDetailsResponse extends ResponseBarrel {
-    description?: string;
-    componentSections: ZeplinComponentSection[];
-    sections?: ResponseScreenSection[];
-}
