+++ src/common/domain/barrel/data/OrganizationStyleguidesStore.ts
@@ -1,9 +1,9 @@
-import { getOrganizationStyleguides } from "../../../common/domain/api/api";
-import BarrelType from "../../../common/domain/barrel/BarrelType";
+import { getOrganizationStyleguides } from "../../api/api";
+import BarrelType from "../BarrelType";
 import BarrelStore from "./BarrelStore";
 import ResponseBarrel from "../model/ResponseBarrel";
 import StyleguidesResponse from "../model/StyleguidesResponse";
-import BaseError from "../../../common/domain/error/BaseError";
+import BaseError from "../../error/BaseError";
 
 export default class OrganizationStyleguidesStore extends BarrelStore<StyleguidesResponse> {
     protected type = BarrelType.Styleguide;
