+++ src/common/domain/barrel/data/PersonalStyleguidesStore.ts
@@ -1,5 +1,5 @@
-import { getPersonalStyleguides } from "../../../common/domain/api/api";
-import BarrelType from "../../../common/domain/barrel/BarrelType";
+import { getPersonalStyleguides } from "../../api/api";
+import BarrelType from "../BarrelType";
 import BarrelStore from "./BarrelStore";
 import ResponseBarrel from "../model/ResponseBarrel";
 import StyleguidesResponse from "../model/StyleguidesResponse";
