+++ src/common/domain/barrel/model/BarrelDetailsResponse.ts
@@ -1,0 +1,9 @@
+import ZeplinComponentSection from "../../zeplinComponent/model/ZeplinComponentSection";
+import ResponseBarrel from "./ResponseBarrel";
+import ResponseScreenSection from "../../screen/model/ResponseScreenSection";
+
+export default interface BarrelDetailsResponse extends ResponseBarrel {
+    description?: string;
+    componentSections: ZeplinComponentSection[];
+    sections?: ResponseScreenSection[];
+}
