+++ src/common/domain/barrel/model/ProjectDetailsResponse.ts
