+++ src/common/domain/barrel/model/StyleguideDetailsResponse.ts
