+++ src/common/domain/barrel/model/StyleguidesResponse.ts
