+++ src/common/domain/screen/model/ResponseScreenSection.ts
