+++ src/common/domain/zeplinComponent/data/BarrelDetailsStoreProvider.ts
@@ -1,6 +1,6 @@
 import BarrelType from "../../barrel/BarrelType";
-import ProjectDetailsStore from "../../../../coco/zeplinComponent/data/ProjectDetailsStore";
-import StyleguideDetailsStore from "../../../../coco/zeplinComponent/data/StyleguideDetailsStore";
+import ProjectDetailsStore from "../../barrel/data/ProjectDetailsStore";
+import StyleguideDetailsStore from "../../barrel/data/StyleguideDetailsStore";
 import Store from "../../store/Store";
 import BarrelDetails from "../model/BarrelDetails";
 import CacheHolder from "../../store/CacheHolder";
