+++ src/session/flow/sessionFlow.ts
@@ -5,8 +5,8 @@ import UriHandler from "../../common/domain/uri/UriHandler";
 import { isUri } from "../../common/vscode/uri/uriUtil";
 import ConfigCodeLensProvider from "../../coco/config/codeLens/ConfigCodeLensProvider";
 import BarrelDetailsStoreProvider from "../../common/domain/zeplinComponent/data/BarrelDetailsStoreProvider";
-import BarrelsStoreProvider from "../../coco/barrel/data/BarrelsStoreProvider";
-import WorkspacesStore from "../../coco/barrel/data/WorkspacesStore";
+import BarrelsStoreProvider from "../../common/domain/barrel/data/BarrelsStoreProvider";
+import WorkspacesStore from "../../common/domain/barrel/data/WorkspacesStore";
 import configuration from "../../common/domain/extension/configuration";
 import { showNotLoggedInError } from "../../common/domain/error/errorUi";
 import MessageBuilder from "../../common/vscode/message/MessageBuilder";
